# docs/swagger.yaml
definitions:
  auth_controller.SignInInput:
    properties:
      password:
        maxLength: 50
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 8
        type: string
    required:
    - password
    - username
    type: object
  auth_controller.SignInOutput:
    properties:
      exp_time:
        type: integer
      token:
        type: string
    type: object
  auth_controller.SignUpInput:
    properties:
      password:
        maxLength: 50
        minLength: 8
        type: string
      username:
        maxLength: 50
        minLength: 8
        type: string
    required:
    - password
    - username
    type: object
  auth_controller.SignUpOutput:
    properties:
      success:
        type: boolean
    type: object
...
info:
  contact: {}
  description: Hung.Phan apply for junior Golang developer position at EMVN
  title: EMVN API
  version: "1.0"
paths:
  /auth/signin:
    post:
      consumes:
      - application/json
      description: Sign in user, return token and exp time
      parameters:
      - description: Sign in user
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth_controller.SignInInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth_controller.SignInOutput'
      summary: Sign in user
      tags:
      - Auth
  /auth/signup:
    post:
      consumes:
      - application/json
      description: Sign up new user
      parameters:
      - description: Sign up new user
        in: body
        name: body
        required: true
        schema:
          $ref: '#/definitions/auth_controller.SignUpInput'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/auth_controller.SignUpOutput'
      summary: Sign up new user
      tags:
      - Auth
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"